Rename StatusItem → AutoStart
