+++ bankid-idp/bankid-idp-frontend/src/components/AutoStart.vue
